requirements.txt
Flask
Flask-SQLAlchemy
Flask-Babel
Flask-Login
Flask-WTF
Pillow
